extensions/positron-python/python_files/posit/test-requirements.txt
fastcore
geopandas
haystack-ai
holoviews
hvplot
ibis-framework; python_version == '3.9'
ibis-framework[duckdb]; python_version >= '3.10'
ipykernel
ipython
ipywidgets
lightning
matplotlib
numpy
pandas
plotly
plotnine
polars
polars[timezone]; sys_platform == 'win32'
pyarrow
pytest
pytest-asyncio
pytest-mock
syrupy
torch
scipy
snowflake-connector-python
sqlalchemy

# putting this last because holoviews is picky about dependency versions (including bokeh),
# so uv's resolver needs to tackle holoviews first
bokeh
